+++ src/libs/utils/time/time.cpp
@@ -665,7 +665,7 @@ Time::str(bool utc) const
 #ifdef __FreeBSD__
     snprintf(__timestr, TIMESTR_SIZE, "%i:%li", __time.tv_sec, __time.tv_usec);
 #else
-    snprintf(__timestr, TIMESTR_SIZE, "%li:%li", __time.tv_sec, __time.tv_usec);
+    snprintf(__timestr, TIMESTR_SIZE, "%li:%li", __time.tv_sec, (long)__time.tv_usec);
 #endif
   } else {
     tm time_tm;
@@ -695,7 +695,7 @@ Time::str_r(char *s, bool utc)
 #ifdef __FreeBSD__
     snprintf(s, TIMESTR_SIZE, "%i:%li", __time.tv_sec, __time.tv_usec);
 #else
-    snprintf(s, TIMESTR_SIZE, "%li:%li", __time.tv_sec, __time.tv_usec);
+    snprintf(s, TIMESTR_SIZE, "%li:%li", __time.tv_sec, (long)__time.tv_usec);
 #endif
   } else {
     tm time_tm;
